Advent of Code 2021 day 2 done

+++ aoc2021/algos.py
@@ -1,4 +1,4 @@
-def sliding_window(nums: list[int], start: int, lookback: int) -> float:
+def sliding_window(nums: list[int], start: int, lookback: int) -> int:
     total = 0
     for i in range(start - (lookback - 1), start + 1):
         total += nums[i]
